feat(lib): replace iuab_compiler_free() with iuab_compiler_fini()
This new function doesn't call free() on the iuab_compiler, so it can
also be used to clean up compilers that weren't dynamically allocated
using malloc().

BREAKING CHANGE: Calls to iuab_compiler_free() must be replaced with a
call to iuab_compiler_fini() followed by a call to free() on the
iuab_compiler struct.